+++ include/TiroTucano.hpp
@@ -5,7 +5,7 @@
 
 class TiroTucano : public Projetil {
     public:
-        TiroTucano(float pos_x, float pos_y);
+        TiroTucano(double pos_x, double pos_y);
         ~TiroTucano();
 
         void update() override;
